refactor(frontend): updated api file structure
