+++ new/src/frontend/acc2/src/api/compute/types.ts
@@ -1,0 +1,7 @@
+export type ComputeResponse = {
+  molecules: string[];
+  configs: {
+    method: string;
+    parameters: string;
+  }[];
+};
