+++ new/src/frontend/acc2/src/api/methods/types.ts
@@ -1,0 +1,6 @@
+export type SuitableMethods = {
+  methods: string[];
+  parameters: {
+    [key: string]: string[];
+  };
+};
