+++ new/src/frontend/acc2/src/hooks/mutations/use-available-parameters-mutation.ts
@@ -1,4 +1,4 @@
-import { getAvailableParameters } from "@acc2/api/parameters";
+import { getAvailableParameters } from "@acc2/api/parameters/parameters";
 import { useMutation } from "@tanstack/react-query";
 
 export const useAvailableParametersMutation = () => {
