+++ new/src/frontend/acc2/src/hooks/mutations/use-computation-setup-mutation.ts
@@ -1,4 +1,4 @@
-import { setup } from "@acc2/api/compute";
+import { setup } from "@acc2/api/compute/compute";
 import { useMutation } from "@tanstack/react-query";
 
 export const useComputationSetupMutation = () => {
