+++ new/src/frontend/acc2/src/hooks/mutations/use-molecules-mutation.ts
@@ -1,4 +1,4 @@
-import { getMolecules } from "@acc2/api/compute";
+import { getMolecules } from "@acc2/api/compute/compute";
 import { useMutation } from "@tanstack/react-query";
 
 export const useMoleculesMutation = () => {
