+++ new/src/frontend/acc2/src/hooks/queries/use-suitable-methods-query.ts
@@ -1,4 +1,4 @@
-import { getSuitableMethods } from "@acc2/api/methods";
+import { getSuitableMethods } from "@acc2/api/methods/methods";
 import { useQuery } from "@tanstack/react-query";
 
 export const useSuitableMethodsQuery = (computationId: string) => {
